requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
soupsieve==2.5
urllib3==2.2.1
